// src/oncobridge-web/src/app/api/models/temporal-occurrence-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PartialDateResponse } from '../models/partial-date-response';
import { PartialPeriodResponse } from '../models/partial-period-response';
export interface TemporalOccurrenceResponse {
  date?: (null | PartialDateResponse);
  kind: string;
  period?: (null | PartialPeriodResponse);
}